requirements.txt
# --- Core functionality ---
streamlit==1.35.0
sentence-transformers==2.7.0
scikit-learn==1.4.2

# --- Resume file handling ---
pdfminer.six==20221105
python-docx==1.1.0
docx2txt==0.8
PyMuPDF==1.22.0

# --- NLP: resume parsing & skills extraction ---
nltk==3.8.1
spacy==3.7.4
pyresparser==1.0.6

# --- Summarization models ---
transformers==4.41.1
torch==2.3.0

# --- Visualization (Optional) ---
plotly==5.22.0
matplotlib==3.9.0
altair==5.3.0

# --- Utility ---
pandas==2.2.2
numpy==1.26.4



streamlit
sentence-transformers
transformers
torch
scikit-learn
pdfplumber
python-docx
spacy
keybert
